refactor(052): name the neutral modules for what they hold
`neutral` named these modules by contrast with the per-runtime modules they
displace, not by what they contain — both declarations said as much ("the
runtime-independent replacement for the two client modules"). Wave B deletes
that contrast partner, after which `neutral` distinguishes the module from
nothing: everything left in each crate is runtime-neutral, as `tunnel.rs`
already was without needing the word in its name.

The decay had started: in `aimdb-knx-connector/src/lib.rs`, `pub mod neutral`
sat under a `// Platform-specific implementations` header saying the opposite
of what it is.

Renaming now because `pub mod neutral` is public API. Wave A is additive and
unreleased, so this is the last moment the change is free rather than
breaking.

  aimdb-knx-connector/src/neutral.rs    -> src/client.rs
  aimdb-serial-connector/src/neutral.rs -> src/framer.rs
  aimdb-tcp-connector/tests/neutral_pool.rs   -> tests/accept_pool.rs
  aimdb-serial-connector/tests/neutral_framed.rs -> tests/framed.rs
  aimdb-embassy-adapter/tests/neutral_udp.rs  -> tests/udp.rs

This also settles on the convention core and the adapters already used for
the same layer — `session::io` and `net`, both named for their contents.

Pure rename: no content changes beyond the module declarations, one import,
the misfiled header comment, the Makefile's `--test` target, and the
changelog and design-doc pointers into these paths. Prose uses of the word
where it is a genuine adjective ("runtime-neutral", "role-neutral") are
untouched.

Verified: clippy `-D warnings` on both connectors for tokio and the
`thumbv7em` embassy cross-compile; knx (37 lib + 35 integration), serial,
`accept_pool` over two real embassy stacks, `udp`; `make doc`;
`cargo fmt --check`.
