+++ aimdb-knx-connector/src/lib.rs
@@ -149,12 +149,13 @@ pub use knx_pico::dpt::{Dpt1, Dpt5, Dpt9, DptDecode, DptEncode};
 // Runtime-neutral KNX/IP tunneling state machine shared by both transports.
 pub mod tunnel;
 
-// Platform-specific implementations
-// One connection task generic over core's neutral datagram traits — the
-// runtime-independent replacement for the two client modules.
+// The connection task: one body for both runtimes, generic over core's
+// datagram and delay traits. Supersedes the two per-runtime client modules
+// below, which it will replace outright.
 #[cfg(any(feature = "tokio-runtime", feature = "embassy-runtime"))]
-pub mod neutral;
+pub mod client;
 
+// Platform-specific implementations
 #[cfg(feature = "tokio-runtime")]
 pub mod tokio_client;
