+++ aimdb-serial-connector/src/framer.rs
