+++ aimdb-tcp-connector/Cargo.toml
@@ -52,7 +52,7 @@ _test-tokio = ["tokio-runtime", "dep:aimdb-tokio-adapter"]
 # the `_test-tokio` build too). Run with `--features _test-embassy-loopback`.
 _test-embassy-loopback = [
     "embassy-runtime",
-    # The adapter's neutral transports, exercised by `tests/neutral_pool.rs`
+    # The adapter's neutral transports, exercised by `tests/accept_pool.rs`
     # over the same two real stacks as `embassy_loopback.rs`.
     "aimdb-embassy-adapter/net",
     "embassy-net/medium-ip",
